feat: disable changelog writing
While developing our wrapper around auto-changelog, I discovered that a recent
change to the tool would make it difficult to generate changelogs for our specific
use case. Our requirements for changelogs specify the ability to ignore pre-release
commits until reaching a full release. The tools that I've tested assume that
you don't care whether a change was included in a pre-release or full release
as far as your changelog is concerned.

In order to support our requirements, we will create a simple tool to parse
commits from `git log` directly and handle them in the way we see fit. For
now, we're disabling any writeouts to changelogs via auto-changelog, and
providing it's output simply to stdout in case it's useful in the interim.